+++ wbooks/src/app/screens/Login/index.tsx
@@ -4,6 +4,8 @@ import { useForm, Controller } from 'react-hook-form';
 import { useNavigation } from '@react-navigation/native';
 import bcInicio from '@assets/bc_inicio.png';
 import wbooksLogo from '@assets/wbooks_logo.png';
+import actionCreators from '@redux/login/actions';
+import { useDispatch, useSelector } from 'react-redux';
 
 import styles from './styles';
 
@@ -13,10 +15,13 @@ function Login() {
   const EMAIL_REGEX = /^(([^<>()\]\\.,;:\s@"]+(\.[^<>()\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{1,}))$/;
   const PASSWORD_REGEX = /^([A-Z]|[0-9]){5,}$/i;
 
+  const dispatch = useDispatch();
+
   const navigation = useNavigation();
-  const onSubmit = () => {
-    navigation.navigate('Library');
+  const onSubmit = (data: any) => {
+    dispatch(actionCreators.login(data.user, data.password, navigation));
   };
+  const user = useSelector((state: any) => state.login);
 
   return (
     <ImageBackground source={bcInicio} style={styles.imageBackground}>
@@ -72,6 +77,7 @@ function Login() {
       <TouchableOpacity style={styles.buttonLogin} onPress={handleSubmit(onSubmit)}>
         <Text style={styles.textButton}>INGRESAR</Text>
       </TouchableOpacity>
+      {user.userError && <Text style={styles.textError}>{user.userError}</Text>}
       <Text style={styles.footer}>Designed, developed and used by woloxers</Text>
     </ImageBackground>
   );
