+++ wbooks/src/redux/book/actions.ts
@@ -1,13 +1,11 @@
-import { createTypes, completeTypes, withPostSuccess } from 'redux-recompose';
-import { bookService, login } from '@services/BooksService';
-import { Dispatch } from 'redux';
-import { ApiOkResponse } from 'apisauce';
+import { createTypes, completeTypes } from 'redux-recompose';
+import { bookService } from '@services/BooksService';
 
-export const actions = createTypes(completeTypes(['GET_BOOKS', 'LOGIN'], ['SET_SEARCH']), '@@BOOKS');
+export const actions = createTypes(completeTypes(['GET_BOOKS'], ['SET_SEARCH']), '@@BOOKS');
 
 export const targets = { books: 'books', search: 'search' };
 
-export const actionsBook = {
+const actionCreators = {
   getBooks: (headers: any) => ({
     type: actions.GET_BOOKS,
     target: targets.books,
@@ -18,20 +16,6 @@ export const actionsBook = {
     type: actions.SET_SEARCH,
     payload: search,
     target: targets.search
-  })
-};
-
-const actionCreators = {
-  login: (email: string, password: string) => ({
-    type: actions.LOGIN,
-    service: login,
-    payload: { email, password },
-    injections: [
-      withPostSuccess((dispatch: Dispatch<any>, response: ApiOkResponse<any>) => {
-        const { headers } = response;
-        dispatch(actionsBook.getBooks(headers));
-      })
-    ]
   })
 };
